+++ src/transform-sdk/rust/Cargo.toml
@@ -1,18 +1,23 @@
-[package]
-name = "redpanda-transform-sdk"
+[workspace]
+resolver = "2"
+members = [
+  "core",
+  "core-sys",
+  "core-types",
+]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2021"
-description = "A Rust library for authoring Redpanda Data Transforms."
+rust-version = "1.72.0"
+authors = ["Redpanda Data Transforms Team"]
 homepage = "https://redpanda.com"
 repository = "https://github.com/redpanda-data/redpanda"
 license = "Apache-2.0"
 keywords = ["redpanda", "wasm", "kafka", "data-transform"]
 categories = ["wasm"]
 
-[dev-dependencies]
-quickcheck = "1"
+[workspace.dependencies]
+quickcheck = "1.0"
 rand = "0.8"
-anyhow = "1.0.75"
-serde_json = "1.0.108"
-apache-avro = "0.16.0"
-
+anyhow = "1.0"
