+++ src/transform-sdk/rust/core-types/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "redpanda-transform-sdk-types"
+description = "Types for redpanda-transform-sdk"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
