+++ src/transform-sdk/rust/core/examples/json2avro.rs
